refactor: break new circular dependencies from evo snapshot work
